MULE-7490 - Removing FilePersistenceTestCase which was failing because depends on the old way to store queue files. This test is not necessary because the same is tested in several other test cases - Only calling delete on temporary folder for working directory when it has a root if not fails when an assumption is false since for some reason junit do not calls the @before methods but calls the @after methods

+++ core/src/test/java/org/mule/tck/junit4/AbstractMuleContextTestCase.java
@@ -281,7 +281,12 @@ public abstract class AbstractMuleContextTestCase extends AbstractMuleTestCase
             disposeContext();
             doTearDownAfterMuleContextDispose();
         }
-        workingDirectory.delete();
+
+        //When an Assumption fails then junit doesn't call @Before methods so we need to avoid executing delete if there's no root folder.
+        if (workingDirectory.getRoot() != null)
+        {
+            workingDirectory.delete();
+        }
     }
 
     protected void doTearDownAfterMuleContextDispose() throws Exception
